fixed some leaks, fixed some C99 standard items (remove size_t)

+++ buffer.h
@@ -2,14 +2,14 @@
 #define STRING_UTILS_H_
 
 typedef struct {
-  size_t length;
-  size_t capacity;
-  size_t expand_length;
+  unsigned int length;
+  unsigned int capacity;
+  unsigned int expand_length;
   char * content;
 } buffer_data;
 
-buffer_data * buffer_init(const size_t initial_size);
-ssize_t buffer_append(buffer_data *buffer, const char *append_data);
+buffer_data * buffer_init(const unsigned int initial_size);
+int buffer_append(buffer_data *buffer, const char *append_data);
 void buffer_destroy(buffer_data * buffer);
 
 #endif // STRING_UTILS_H_
